% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarised_experiments.R
\name{annotate_se}
\alias{annotate_se}
\title{Update SummarisedExperiment}
\usage{
annotate_se(
  summarised_experiment,
  sample_metadata,
  multiqc_path = NULL,
  gtf = FALSE,
  qc = TRUE,
  mode = "replace"
)
}
\arguments{
\item{summarised_experiment}{the SummarisedExperiment to update}

\item{sample_metadata}{a data.frame to replace or amend colData}

\item{multiqc_path}{the path to a multiqc report from which to extract the QC metrics or gtf}

\item{gtf}{a logical indicating whether to extract the gtf file from the multiqc report and replace gene annotations for rowData}

\item{qc}{a logical indicating whether to extract the QC metrics from the multiqc report and replace or amend gene annotations to rowData}

\item{mode}{one of c("replace", "amend"), should the provided data replace or amend the existing elements? Default: "replace"}
}
\value{
a SummarisedExperiment
}
\description{
Updates or replaces the \code{colData} or \code{rowData} of a SummarisedExperiment.
}
\details{
Genes and associated metadata come from the \code{gtf} logged in the multiqc report.

Sample metadata is matched by the \code{sample} ID from the \code{sample_metadata} table and the
QC metrics from the pipeline run. If the multiqc path refers to a full QC run (\code{qc = TRUE}
set in \code{\link[=run_nf_core_rnaseq]{run_nf_core_rnaseq()}}, the QC metrics come from
\code{{out_path}/multiqc/star_salmon/multiqc_data/multiqc_general_stats.txt}. If a Salmon-only
run (\code{qc = FALSE} set in \code{\link[=run_nf_core_rnaseq]{run_nf_core_rnaseq()}}), the QC metrics come from
\code{{out_path}/multiqc/multiqc_data/multiqc_salmon.txt}.
}
\author{
Denis O'Meally
}
