% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cml_mrna_2021.R
\docType{data}
\name{cml_mrna_2021_GRCm38_HLT}
\alias{cml_mrna_2021_GRCm38_HLT}
\title{SummarisedExperiment of 75 samples}
\format{
A \code{SummarisedExperiment} object containing 226 samples.
\describe{
\item{counts}{\code{mat} Salmon \emph{transcript length bias corrected} gene counts}
\item{abundance}{\code{mat} Salmon \emph{transcript length bias corrected} TPM}
}
}
\source{
\url{http://cgt.coh.org/MHO/CML.mRNA.2021/results/rnaseq3.5.GRCm38.HLT/salmon/salmon.merged.gene_counts_length_scaled.rds}
}
\usage{
data(cml_mrna_2021_GRCm38_HLT)
}
\description{
A \code{SummarisedExperiment} generated by the nf-core/rnaseq v3.5 pipeline using Salmon
and tximport. See the pipeline \href{https://nf-co.re/rnaseq/3.5/output#pseudo-alignment-and-quantification}{documentation}
for details. The \href{../articles/genomes.html}{reference}
is GRCm38 using Ensembl annotation v81 (July 2015) with human leukemic transcripts (MYH11, BCR & ABL1).
}
\details{
Only the Salmon \emph{\href{https://nf-co.re/rnaseq/3.5/output#pseudo-alignment-and-quantification}{transcript length bias corrected}} counts are available in this package;
for other datasets, see the \href{http://cgt.coh.org/MHO/CML.mRNA.2021/results/rnaseq3.5.GRCm38.HLT/}{full pipeline output}
}
\keyword{2021}
\keyword{cml}
\keyword{mrna}
