% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cml_mrna.R
\docType{data}
\name{cml_mrna_GRCm38_HLT}
\alias{cml_mrna_GRCm38_HLT}
\title{Expression matrix of 226 samples}
\format{
A \code{SummarisedExperiment} object containing 226 samples.
\describe{
\item{counts}{\code{mat} Salmon \emph{transcript length bias corrected} gene counts}
\item{abundance}{\code{mat} Salmon \emph{transcript length bias corrected} TPM}
}
}
\source{
\url{merge_mrna.html}
}
\usage{
data(cml_mrna_GRCm38_HLT)
}
\description{
A \code{SummarisedExperiment} generated by \code{cbind}ing the \code{cml_mrna_2021} and \code{cml_mrna_2022} datasets.
After QC, the following samples were removed:
\code{c("F545.7", "X490.17")}
}
\details{
The original datasets were prepared using the nf-core/rnaseq v3.5 pipeline with Salmon
and tximport. See the pipeline \href{https://nf-co.re/rnaseq/3.5/output#pseudo-alignment-and-quantification}{documentation}
for details. The \href{../articles/genomes.html}{reference}
is GRCm38 using Ensembl annotation v81 (July 2015) with human leukemic transcripts (MYH11, BCR & ABL1).

Only the Salmon \emph{\href{https://nf-co.re/rnaseq/3.5/output#pseudo-alignment-and-quantification}{transcript length bias corrected}} counts are available in this package;
for other datasets, see the full pipeline output for \href{http://cgt.coh.org/MHO/CML.mRNA.2021/results/rnaseq3.5.GRCm38.HLT/}{2021}
and \href{http://cgt.coh.org/MHO/CML.mRNA.2021/results/rnaseq3.5.GRCm38.HLT/}{2022}.
}
\keyword{2021}
\keyword{2022}
\keyword{cml}
\keyword{mrna}
