% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{find_elbow}
\alias{find_elbow}
\title{find_elbow}
\usage{
find_elbow(seurat_object, plot = FALSE)
}
\arguments{
\item{seurat_object}{a Seurat object}

\item{plot}{TRUE/FALSE to make a plot of the elbow, Default = FALSE}
}
\value{
either an \code{integer} or a \code{ggplot2} object.
}
\description{
Find elbow in PCA for Seurat objects.
}
\details{
Inspired by
https://hbctraining.github.io/scRNA-seq/lessons/elbow_plot_metric.html
Determine percent of variation associated with each PC,
Calculate cumulative percent variation for each PC,
Determine which PC exhibits cumulative variation greater than 90\%,
and \% variation associated with the PC is less than 5\%, then
Choose the minimum of the two

Optionally makes a plot of the cumulative percent variation.
}
\author{
Denis O'Meally
}
