% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarised_experiments.R
\name{find_outliers_se}
\alias{find_outliers_se}
\title{Run OUTRIDER pipeline on a SummarizedExperiment}
\usage{
find_outliers_se(summarised_experiment)
}
\arguments{
\item{summarised_experiment}{a SummarisedExperiment to run the OUTRIDER pipeline on}
}
\value{
a list of plots and tables
}
\description{
Runs the OUTRIDER pipeline on a summarised experiment, identifying outlying samples and genes.
}
\author{
Denis O'Meally
}
