% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_data.R
\name{get_latest_pin_version}
\alias{get_latest_pin_version}
\title{Get latest pin version}
\usage{
get_latest_pin_version(pin_name)
}
\arguments{
\item{pin_name}{the name of the pin for which the version is required.}
}
\value{
the pin version
}
\description{
Finds a pin with a given name and returns the latest
version from the pin_board. The pinboard must be set up for the current
R session using the \code{\link[=use_pinboard]{use_pinboard()}} function.
}
\details{
Use this function to get the latest version of a pin.
Uses the \code{\link[pins:pin_versions]{pins::pin_versions()}} function to get the list of pin
versions and \code{dplyr} to get the top element of the returned list.
}
\examples{
\dontrun{
if (interactive()) {
    metadata_mmu_version <- get_latest_pin_version("metadata_mmu.csv")
}
}
}
\seealso{
\code{\link[pins]{pin_versions}}
\code{\link[dplyr]{pull}}, \code{\link[dplyr]{nth}}
}
