% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_data.R
\name{get_pin_list}
\alias{get_pin_list}
\title{Get pin list}
\usage{
get_pin_list()
}
\value{
list of pin names
}
\description{
Return a list of pin names from the pin_board.
If \code{onedrive} is TRUE, the pin is retrieved from Sharepoint.
If \code{onedrive} is FALSE, the pin is retrieved from the MHO
haemdata folder.
}
\examples{
\dontrun{
if (interactive()) {
    get_pin_list()
}
}
}
\seealso{
\code{\link[pins]{pin_list}}
}
