% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rnaseq_se.R
\name{get_rnaseq_se}
\alias{get_rnaseq_se}
\title{get_rnaseq_se}
\usage{
get_rnaseq_se(isilon_path)
}
\arguments{
\item{isilon_path}{a path on the COH Isilon store for raw data. Should begin with \verb{/net}.}
}
\value{
a \code{SummarisedExperiment} object
}
\description{
Get a \code{SummarisedExperiment} object from an nf-core/rnaseq run
}
\details{
Pulls in a \code{SummarisedExperiment} from the nf-core/rnaseq Salmon folder, and annotates
experiment metadata derived from the path provided.
}
\author{
Denis O'Meally
}
