% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haemdata-packages.R
\docType{package}
\name{haemdata}
\alias{haemdata}
\title{haemdata: PSON omics data}
\description{
Harmonised data from mouse RNAseq projects processed with \href{https://nf-co.re}{nf-core} pipelines
and published as the R package \code{haemdata} on GitHub.
}
