% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haemdata.R
\docType{data}
\name{haemdata_env}
\alias{haemdata_env}
\title{Package environment}
\format{
An object of class \code{environment} of length 3.
}
\usage{
haemdata_env
}
\description{
Package environment
}
\details{
The \code{haemdata_env} package environment holds variables and objects used
by the package, for example the \code{pin_board}. See see the \href{https://github.com/cohmathonc/haemdata/blob/main/R/haemdata.R}{package source}
for more context.
}
\keyword{datasets}
