% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smrnaseq.R
\name{make_isomir}
\alias{make_isomir}
\title{make_isomir}
\usage{
make_isomir(mirtop_isomirs, metadata)
}
\arguments{
\item{mirtop_isomirs}{a data.frame containing the mirtop isomiR table}

\item{metadata}{a data.frame containing the colData to bind}
}
\value{
an IsomirDataSeq object (a subclass of SummarizedExperiment)
}
\description{
Make an isomiRs object from the mirtop isomir table and a metadata table
}
\details{
Given a mirtop isomir table and a metadata table, this will return an
\href{https://www.bioconductor.org/packages/release/bioc/html/isomiRs.html}{\code{isomiRs}}
SummarizedExperiment object.
}
\author{
Denis O'Meally
}
