% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_logo}
\alias{make_logo}
\title{Make the package logo}
\usage{
make_logo(url = NULL, text = NULL)
}
\arguments{
\item{url}{a URL to a picture}

\item{text}{the name on the sticker}
}
\value{
a hex sticker
}
\description{
makes a hex sticker given a url
}
