% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_metadata.R
\name{make_metadata_hsa}
\alias{make_metadata_hsa}
\title{Make minimal metadata for HSA}
\usage{
make_metadata_hsa(sample_sheet)
}
\arguments{
\item{sample_sheet}{A data.frame with a sample sheet}
}
\value{
a data.frame sorted by batch, patient_id, date
}
\description{
Collects minimal metadata for human samples from the "sample summary"
excel sheet for FLT3 samples from COH Biobank and for MDS patients from the
EGA (\href{https://ega-archive.org/datasets/EGAD00001003891}{EGAD00001003891}).
}
\details{
The resulting table is sorted by \code{batch}, \code{patient_id} and \code{date}.
}
\author{
Denis O'Meally
}
