% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_metadata.R
\name{make_metadata_mmu}
\alias{make_metadata_mmu}
\title{Make metadata for mouse AML samples}
\usage{
make_metadata_mmu(sample_sheet_all_mice)
}
\arguments{
\item{sample_sheet_all_mice}{a data.frame produced by row binding all mouse sample sheets
detailed in \href{https://github.com/drejom/haemdata/blob/main/R/import_metadata.R}{R/import_metadata.R}.}
}
\value{
a data.frame
}
\description{
This function prepares the \code{metadata_mmu} object for all RNAseq libraries from AML and CML mice.
Minimal metadata fields include sample, fastq_1, fastq_2, strandedness,
mouse_id, tissue, week, timepoint, batch, treatment, genotype, sex, dob, project.
}
\details{
Raise an \href{https://github.com/drejom/haemdata/issues}{issue on GitHub}
to report erroneous or missing records.

\code{weeks} are read in from \code{data-raw/timepoints.xlsx} and \code{left_joined} to the
\code{sample_sheet_all_mice} by \code{timepoint_project}. For samples with the time point \code{L} (leukemia) or \code{END},
\code{weeks} is set to 1 + the penultimate sample for a mouse. Bone marrow samples are assigned the
\code{timepoint} \code{NA} and \code{weeks} is also set to 1 + the penultimate sample for that mouse.
}
\author{
Denis O'Meally
}
