% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_minimal_metadata_hsa.R
\name{make_minimal_metadata_hsa}
\alias{make_minimal_metadata_hsa}
\title{Make minimal metadata for HSA}
\usage{
make_minimal_metadata_hsa()
}
\value{
a data.frame
}
\description{
Collects minimal metadata for human samples from the "sample summary"
excel sheet for FLT3 samples from COH Biobank.
}
\author{
Denis O'Meally
}
