% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_minimal_metadata_mmu.R
\name{make_minimal_metadata_mmu}
\alias{make_minimal_metadata_mmu}
\title{Make metadata for mouse AML samples}
\usage{
make_minimal_metadata_mmu(sample_sheet_all_mice)
}
\arguments{
\item{sample_sheet_all_mice}{a data.frame produced by row binding all mouse sample sheets
derived from \href{https://github.com/drejom/haemdata/blob/main/R/parse_metadata.R}{R/parse_metadata.R}.}
}
\value{
a data.frame
}
\description{
This function prepares the \code{minimal_metadata_mmu} object for all RNAseq libraries from AML and CML mice.
Minimal metadata fields include sample, fastq_1, fastq_2, strandedness,
mouse_id, tissue, timepoint, batch, treatment, genotype, sex, dob, project.
}
\details{
Raise an \href{https://github.com/drejom/haemdata/issues}{issue on GitHub}
to report erroneous or missing records.
}
\author{
Denis O'Meally
}
