% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarised_experiments.R
\name{make_tpm_matrix}
\alias{make_tpm_matrix}
\title{Write an expression matrix to file}
\usage{
make_tpm_matrix(summarised_experiment, drop = NULL, tpm = 1, samples = 5)
}
\arguments{
\item{summarised_experiment}{a \code{SummarisedExperiment} object}

\item{drop}{samples (columns) to remove}

\item{tpm}{minimum TPM for a gene to be kept, Default: 1}

\item{samples}{minimum number of samples a gene must be present in to be kept, Default: 5}
}
\value{
a list
\itemize{
\item \code{tpm_matrix} an expression matrix, with genes as rows nad samples as columns
\item \code{name} a name indicating the \code{tpm} and \code{sample} thresholds used
}
}
\description{
Write an TPM expression matrix to a CSV file in \code{extdata}
}
\details{
This function makes an expression matrix, filtering out genes with
fewer than \code{tpm} reads in at least \code{samples} samples.
}
\examples{
\dontrun{
if (interactive()) {
    make_tpm_matrix(cml_mrna_GRCm38_HLT, c("F545.7", "X490.17"), tpm = 1, samples = 5)
}
}
}
\seealso{
\code{\link[SummarizedExperiment]{SummarizedExperiment-class}}
\code{\link[utils]{write.table}}
}
