% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_mrna.R
\name{merge_mrna}
\alias{merge_mrna}
\title{Merge a pair of \code{SummarisedExperiments}}
\usage{
merge_mrna(summarised_experiment1, summarised_experiment2, drop = NULL)
}
\arguments{
\item{summarised_experiment1, }{name of a \code{SummarisedExperiment} object to merge}

\item{summarised_experiment2, }{name of the other \code{SummarisedExperiment} object to merge}

\item{drop}{a vector of sample names to remove from the merged \code{SummarisedExperiment}}
}
\value{
a \code{SummarisedExperiment} object
}
\description{
Takes a pair of \code{SummarisedExperiments} produced by \code{\link[=get_rnaseq_se]{get_rnaseq_se()}},
and merges them into a single \code{SummarisedExperiment}
}
\author{
Denis O'Meally
}
