% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarised_experiments.R
\name{merge_mrna_se}
\alias{merge_mrna_se}
\title{Merge a pair of \code{SummarisedExperiments}}
\usage{
merge_mrna_se(se1, se2)
}
\arguments{
\item{se1, }{name of a \code{SummarisedExperiment} object to merge}

\item{se2, }{name of the other \code{SummarisedExperiment} object to merge}
}
\value{
a \code{SummarisedExperiment} object
}
\description{
Takes a pair of \code{SummarisedExperiments} produced by \code{\link[=get_rnaseq_se]{get_rnaseq_se()}},
and merges them into a single \code{SummarisedExperiment}. If the colData differs, we
takes the union of both. Metadata is checked for consistency, and the merge
will not proceed
}
\author{
Denis O'Meally
}
