% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\docType{data}
\name{metadata_hsa}
\alias{metadata_hsa}
\title{Minimal metadata for human samples}
\description{
The make_metadata_hsa() function assembles the metadata for all RNAseq libraries
from patient samples. Minimal metadata fields include library_id, fastq_1, fastq_2, strandedness,
sample_id, tissue, weeks, timepoint, batch, treatment, genotype, sex, dob, project.
}
\details{
For human samples, metadata are sourced from the EGA and supplied excel sheets.

#TODO Describe the studies: MDS & COH Biobank FLT3

Raise an \href{https://github.com/cohmathonc/haemdata/issues}{issue on GitHub}
to report erroneous or missing records.
}
\author{
Denis O'Meally
}
