% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smrnaseq.R
\name{nfcore_mirtop_counts}
\alias{nfcore_mirtop_counts}
\title{nfcore_mirtop_counts}
\usage{
nfcore_mirtop_counts(multiqc_path)
}
\arguments{
\item{multiqc_path}{the full path to a multiqc pipeline report from a nfcore/smrnaseq run
on COH Isilon storage}
}
\value{
a tibble of counts, with genes as rows, and samples as columns
}
\description{
Get the counts table from mirtop output
}
\details{
Given the path of a multiqc pipeline report file, this function will read in \code{mirna.tsv} from the miRTop folder and return a tibble of counts.
}
\author{
Denis O'Meally
}
