% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smrnaseq.R
\name{nfcore_mirtop_isomir}
\alias{nfcore_mirtop_isomir}
\title{nfcore_mirtop_isomir}
\usage{
nfcore_mirtop_isomir(multiqc_path)
}
\arguments{
\item{multiqc_path}{the full path to a multiqc pipeline report from a nfcore/smrnaseq run
on COH Isilon storage}
}
\value{
a data.frame of counts per million, with genes as rows, and samples as columns
}
\description{
Get the isomiRs table from mirtop_rawData.tsv file
}
\details{
Given the path of a multiqc pipeline report file, this function will read in \code{mirtop.tsv}
from the mirtop folder and return a tibble mirtop isomiRs suitable for input to the
\href{https://www.bioconductor.org/packages/release/bioc/html/isomiRs.html}{\code{isomiRs}} package.
}
\author{
Denis O'Meally
}
