% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarised_experiments.R
\name{nfcore_rnaseq_to_se}
\alias{nfcore_rnaseq_to_se}
\title{nf-core/rnaseq to SummarisedExperiment}
\usage{
nfcore_rnaseq_to_se(multiqc_path, sample_metadata)
}
\arguments{
\item{multiqc_path}{path to the multiqc report of a successful nf-core/rnaseq pipeline run}

\item{sample_metadata}{a data.frame to combine with the QC metrics from the pipeline run as colData in the SummarisedExperiment}
}
\value{
a SummarisedExperiment
}
\description{
Makes a SummarisedExperiment from the pipeline run given by \code{multiqc_path}.
}
\details{
Genes and associated metadata come from the \code{gtf} logged in the multiqc report.

Sample metadata is matched by the \code{library_id} from the \code{sample_metadata} table and the
QC metrics from the pipeline run. If the multiqc path refers to a full QC run (\code{qc = TRUE}
set in \code{\link[=run_nf_core_rnaseq]{run_nf_core_rnaseq()}}, the QC metrics come from
\code{{out_path}/multiqc/star_salmon/multiqc_data/multiqc_general_stats.txt}. If a Salmon-only
run (\code{qc = FALSE} set in \code{\link[=run_nf_core_rnaseq]{run_nf_core_rnaseq()}}), the QC metrics come from
\code{{out_path}/multiqc/multiqc_data/multiqc_salmon.txt}.

Assays slots for both \code{counts} and \code{abundance} come from the Salmon output \code{{out_path}/salmon/salmon.merged.gene_counts.rds}
}
\author{
Denis O'Meally
}
