% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smrnaseq.R
\name{nfcore_smrna_cpm}
\alias{nfcore_smrna_cpm}
\title{nfcore_smrna_cpm}
\usage{
nfcore_smrna_cpm(multiqc_path)
}
\arguments{
\item{multiqc_path}{the full path to a multiqc pipeline report from a nfcore/smrnaseq run
on COH Isilon storage}
}
\value{
a data.frame of counts per million, with genes as rows, and samples as columns
}
\description{
Get the counts table from edgeR CPM file
}
\details{
Given the path of a multiqc pipeline report file, this function will read in \code{hairpin_normalized_CPM.txt} from the edger folder and return a tibble of counts per million.
}
\author{
Denis O'Meally
}
