% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smrnaseq.R
\name{nfcore_smrna_qc}
\alias{nfcore_smrna_qc}
\title{nfcore_smrna_qc}
\usage{
nfcore_smrna_qc(multiqc_path)
}
\arguments{
\item{multiqc_path}{the full path to a multiqc pipeline report from a nfcore/smrnaseq run
on COH Isilon storage}
}
\value{
a tibble of counts, with samples as rows, and QC metrics as columns
}
\description{
Get the mirTrace and samtools QC tables
}
\details{
Given the path of a multiqc pipeline report file, this function will read in mirTrace and samtools stats from the multiqc folder and return a tibble.
}
\author{
Denis O'Meally
}
