% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{plot_cohort_survival}
\alias{plot_cohort_survival}
\title{Plot cohort survival}
\usage{
plot_cohort_survival(
  sample_sheet = metadata_mmu_prepub,
  cohort_regex,
  assay_regex = "mRNA"
)
}
\arguments{
\item{sample_sheet}{A data frame containing sample metadata.}

\item{cohort_regex}{A regular expression to filter the cohort of interest.}

\item{assay_regex}{A regular expression to filter the assay of interest.}
}
\value{
A ggplot object showing the effect of treatment on survival for a given cohort and assay.
}
\description{
Plot cohort survival
}
\details{
Uses the \code{survfit} function from the {survival} package to fit a
survival model to the data, and uses the \code{autoplot.survfit} function from
the {ggfortify} package to create a ggplot to visualize the fitted model.
The plot shows the cohort survival over time, with different colors indicating
different treatments.
}
\examples{
\dontrun{
use_pinboard("devel")
get_pin("metadata_mmu.csv") |>
  plot_cohort_survival(cohort_regex = "AML")
}
}
