% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_transgenes.R
\name{plot_transgenes}
\alias{plot_transgenes}
\title{Plot human transgene expression}
\usage{
plot_transgenes(
  summarised_experiment,
  assay = "abundance",
  group_by = "treatment"
)
}
\arguments{
\item{summarised_experiment}{a SummarisedExperiment object}

\item{assay}{the matrix to plot, Default: "abundance"}

\item{group_by}{a string indicating the column to group by}
}
\value{
a ggplot2 plot
}
\description{
Plots expression of genes matching the pattern \verb{HSA_.*_gene} from a SummarisedExpression object.
Log2TPM is values are plotted from the \code{abundance} assay.
}
\author{
Denis O'Meally
}
