% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publish_data.R
\name{publish_metadata}
\alias{publish_metadata}
\title{Publish sample metadata}
\usage{
publish_metadata(metadata)
}
\arguments{
\item{metadata}{a data.frame ready to publish}
}
\value{
a data.frame of the pins and version numbers for the published object
}
\description{
Publishes a metadata table to the \code{MHO/haemdata} folder on Isilon storage in \code{.csv} format.
If \code{teams} is \code{TRUE}, a note is posted to the Teams channel.
}
\author{
Denis O'Meally
}
