% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publish_data.R
\name{publish_se}
\alias{publish_se}
\title{Publish a SummarisedExperiment}
\usage{
publish_se(summarised_experiment)
}
\arguments{
\item{summarised_experiment}{a SummarisedExperiment ready to publish}
}
\value{
the path to the published object
}
\description{
Publishes a SummarisedExperiment to the Haemdata Teams channel and to the \code{MHO/haemdata} folder
on the COH Isilon storage (in \code{.rda} format) along with an expression matrix
(in \code{.csv} format). Files are named according to the SummarisedExperiment's
\code{metadata$oject_name} variable
}
\author{
Denis O'Meally
}
