% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{published_pins}
\alias{published_pins}
\title{Datasets published to the Haemdata pin board}
\format{
A \code{data.frame} with the following columns:
\describe{
\item{counts}{\code{mat} Salmon \emph{transcript length bias corrected} gene counts}
\item{abundance}{\code{mat} Salmon \emph{transcript length bias corrected} TPM}
}
}
\source{
\code{\link{R/publish_data.R}}
}
\usage{
data(published_pins)
}
\description{
TODO
}
\keyword{2021}
\keyword{cml}
\keyword{mrna}
