% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarised_experiments.R
\name{qc_filter_se}
\alias{qc_filter_se}
\title{qc_filter_se}
\usage{
qc_filter_se(summarised_experiment, mapping_threshold = 5)
}
\arguments{
\item{summarised_experiment}{a SummarisedExperiment from a "QC" run (\code{qc = TRUE}
set in \code{\link[=run_nf_core_rnaseq]{run_nf_core_rnaseq()}}}

\item{mapping_threshold}{threshold below which samples will be discarded (STAR uniquely-mapped reads). Default: 5 (\%)}
}
\value{
a SummarisedExperiment. The \code{metadata} slot of the SummarisedExperiment is updated with a record of the samples
removed by the filter (\code{qc_removed}).
}
\description{
Remove poorly mapped samples
}
\details{
If the QC metrics contain a \code{uniquely_mapped_percent} column, remove samples below the specified \code{mapping_threshold}.
}
\author{
Denis O'Meally
}
