% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{rotate_umap}
\alias{rotate_umap}
\title{Rotate UMAP}
\usage{
rotate_umap(integrated, x = FALSE, y = FALSE)
}
\arguments{
\item{integrated}{a Seurat object}

\item{x}{flip the x-axis LOGICAL (TRUE or FALSE)}

\item{y}{flip the y-axis LOGICAL (TRUE or FALSE)}
}
\value{
A Seurat object
}
\description{
Makes the 1st cell in a Seurat object plot in the top left quadrant (ie -1,-1)
This is useful for plotting UMAPs in a consistent way across platforms/runs.
The "x" and "y" parameters rotate the axis by 90 degrees.
}
\author{
Denis O'Meally
}
