% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_nfcore_scrnaseq.R
\name{run_nf_core_scrnaseq}
\alias{run_nf_core_scrnaseq}
\alias{run_nfcore_scrnaseq}
\title{Run nf-core/scrnaseq pipeline on Apollo}
\usage{
run_nfcore_scrnaseq(run_folder, sample_sheet)
}
\arguments{
\item{run_folder}{subfolder in which to run nextflow, keeps the pipeline
from overwriting previous runs and enables the \code{-resume} feature of Nextflow.}

\item{sample_sheet}{nf-core sample sheet with the columns \verb{library_id, fastq_1, fastq_2, strandedness} (see \href{https://github.com/drejom/haemdata/blob/HEAD/R/import_metadata.R}{\code{R/import_metadata.R}})}
}
\value{
a path to the run script, or the multiqc report if it exists.
}
\description{
Makes an \code{sbatch} job script for the nf-core/scrnaseq pipeline and submits it to the Apollo cluster.
The pipeline version is specified by the \code{scrnaseq_release} parameter in the \verb{_targets.R} file. The \href{https://nf-co.re/scrnaseq/2.1.0/usage}{default parameters} are used for the pipeline.
}
\details{
Execution of the pipeline is independent of the package build script, allowing asynchronous development.
If no \code{run_folder} exists, it will be created along with a \code{sample_sheet.csv} and \verb{run_\{ref_genome\}.sh},
which is submitted via \code{sbatch}. A successful run is detected by the presence of a a multiqc report.
An existing run script wont be overwritten by this function so it must be deleted manually.
}
\author{
Denis O'Meally
}
