% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarised_experiments.R
\name{pca_se}
\alias{pca_se}
\title{Perform PCA on a SummarisedExperiment}
\usage{
pca_se(summarised_experiment, assay = "abundance", col_by = NULL)
}
\arguments{
\item{summarised_experiment}{a \code{SummarisedExperiment} object}

\item{assay}{the assay to perform PCA on, Default: "abundance"}

\item{col_by}{column name of colData to use for colouring samples in PCA plots. If none is given,
samples will be coloured by the "condition", "treatment" or "sample" column if any exists, in that order. Default: NULL}
}
\value{
a list of ggplot2 plots
}
\description{
Perform PCA on the \code{assay} of a SummarisedExperiment object and make a biplot, pairs plot and
correlation of metadata columns with PCs.
}
\details{
Could be rewritten, but this is a quick and dirty way to get something working using \code{vignette("PCAtools", package = "PCAtools")}.
}
\examples{
\dontrun{
if (interactive()) {
    se_pca(AML.mRNA.2016_qc_se)
}
}
}
\seealso{
\code{\link[SummarizedExperiment]{SummarizedExperiment-class}}
\code{\link[dplyr]{case_when}}, \code{\link[dplyr]{select}}, \code{\link[dplyr]{reexports}}
\code{\link[PCAtools]{pca}}, \code{\link[PCAtools]{findElbowPoint}}, \code{\link[PCAtools]{biplot}}, \code{\link[PCAtools]{pairsplot}}, \code{\link[PCAtools]{eigencorplot}}, \code{\link[PCAtools]{getComponents}}
\code{\link[janitor]{remove_constant}}
}
\author{
Denis O'Meally
}
