% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_annotate_cell_cycle_MMU.R
\name{seurat_annotate_cell_cycle}
\alias{seurat_annotate_cell_cycle}
\title{Annotate cell cycle}
\usage{
seurat_annotate_cell_cycle(integrated)
}
\arguments{
\item{integrated}{A Seurat object to annotate}
}
\value{
A Surat object with cell cycle annotations
}
\description{
Get the mouse cell cycle markers from the package \code{scran}, and annotate the
Seutat object with the cell cycle labels for each cell.
}
\details{
See \url{https://ucdavis-bioinformatics-training.github.io/2019-single-cell-RNA-sequencing-Workshop-UCD_UCSF/scrnaseq_analysis/scRNA_Workshop-PART1.html}
for more information.
}
\author{
Denis O'Meally
}
