% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{seurat_annotate_cell_type}
\alias{seurat_annotate_cell_type}
\title{Annotate mouse cell type}
\source{
https://bioconductor.org/books/release/SingleRBook/
}
\usage{
seurat_annotate_cell_type(seurat_object)
}
\arguments{
\item{seurat_object}{a Seurat object}
}
\value{
a Seurat object with cell type annotations
}
\description{
Uses the SingleR::SingleR() function to label cell types according
to expression profiles from the \href{https://www.immgen.org/}{Immunological Genome Project}
made available via the the \code{{celldex}} package. The Seurat object per-cell metadata
is updated with \code{cell_type} and \code{cell_type_fine} columns.
}
\author{
Denis O'Meally, Yu-Husan Fu
}
