% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_annotate_type.R
\name{seurat_annotate_type}
\alias{seurat_annotate_type}
\title{Annotate mouse bone marrow cell type}
\source{
https://www.frontiersin.org/articles/10.3389/fcell.2021.622519/full
}
\usage{
seurat_annotate_type(seurat_object)
}
\arguments{
\item{seurat_object}{a list of seurat objects}
}
\value{
a seurat object
}
\description{
Use the \code{\link[clustifyr:clustify]{clustifyr::clustify()}} function to label cell types according
to expression profiles identified by Dolgalev & Tikhonova (2021) in a
meta-analysis of five mouse bone marrow scRNAseq datasets.
}
\author{
Denis O'Meally
}
