% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_annotate_umap_and_clusters.R
\name{seurat_annotate_umap_and_clusters}
\alias{seurat_annotate_umap_and_clusters}
\title{Annotate UMAP and knn clusters}
\usage{
seurat_annotate_umap_and_clusters(seurat_object)
}
\arguments{
\item{seurat_object}{a Seurat object}
}
\value{
a Seurat object with UMAP and knn clusters
}
\description{
Runs the 'standard workflow' using 2000 features and 50 PCs to cluster the data.
}
\author{
Denis O'Meally
}
