% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_integrate_RPCA.R
\name{seurat_integrate_RPCA}
\alias{seurat_integrate_RPCA}
\title{integrate_RPCA}
\usage{
seurat_integrate_RPCA(seurat_object_list)
}
\arguments{
\item{seurat_object_list}{a list of seurat objects}
}
\value{
a seurat object
}
\description{
RPCA integration for Seurat objects
}
\details{
Integrate across orig.ident using RPCA.
https://satijalab.org/seurat/articles/integration_large_datasets.html
}
\author{
Denis O'Meally
}
