% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_integrate_scTransform.R
\name{seurat_sctransform}
\alias{seurat_sctransform}
\title{Integrate a Seurat object using \code{sctransform}}
\usage{
seurat_sctransform(seurat_object)
}
\arguments{
\item{seurat_object}{a Seurat object}
}
\value{
Returns a Seurat object with a new assay (named SCT by default) with counts
being (corrected) counts, data being log1p(counts), scale.data being pearson residuals;
\code{sctransform::vst} intermediate results are saved in misc slot of the new assay.
}
\description{
scTransform integration for Seurat objects
}
\details{
Integrate a Seurat object using \code{sctransform}.
https://satijalab.org/seurat/articles/integration_introduction.html#performing-integration-on-datasets-normalized-with-sctransform-1
}
\author{
Denis O'Meally
}
