% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{write_data}
\alias{write_data}
\title{Write rda file to \verb{data/}}
\usage{
write_data(schemeName, data)
}
\arguments{
\item{schemeName}{the name of the rda file}

\item{data}{the object to save}
}
\description{
This function writes an \code{rda} file in \code{data}
and names it \code{schemeName}. Makes saving objects easier
from within a function. Inspired by \href{https://stackoverflow.com/questions/56293910/create-r-data-with-a-dynamic-variable-name-from-function-for-package}{this}
StackOverflow post.
}
