% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_data.R
\name{get_pin}
\alias{get_pin}
\title{Get pin}
\usage{
get_pin(pin_name, version = NULL)
}
\arguments{
\item{pin_name}{The name of the pin to retrieve}

\item{version}{The version of the pin to retrieve; if NULL, the latest version is used. Default: NULL}
}
\value{
An object from the pin_board
}
\description{
Get an object from the pin_board
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
    metadata_mmu <- get_pin("metadata_mmu.csv")
}
}
}
\seealso{
\code{\link[pins]{pin_read}}
}
