% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_pin_list}
\alias{get_pin_list}
\title{get_pin_list}
\usage{
get_pin_list()
}
\value{
a vector of pin names
}
\description{
Get names of all available Heamdata pin_board pins
}
\details{
The \code{pin_board} is set at the top of in \verb{_targets.R}
}
\examples{
\dontrun{
if (interactive()) {
    get_pin_list()
}
}
}
\seealso{
\code{\link[pins]{pin_list}}
}
