% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarised_experiments.R
\name{get_rnaseq_se}
\alias{get_rnaseq_se}
\title{Get a \code{SummarisedExperiment} object from an nf-core/rnaseq run}
\usage{
get_rnaseq_se(multiqc_path, gtf = TRUE)
}
\arguments{
\item{multiqc_path}{path to the multiqc report of a successful nf-core/rnaseq pipeline run}

\item{gtf}{annotate the SummarisedExperiment using the \code{gtf} recorded in the multiqc report, Default: TRUE}
}
\value{
a \code{SummarisedExperiment} object
}
\description{
Pulls in a \code{SummarisedExperiment} from the nf-core/rnaseq Salmon folder, and annotates
experiment metadata derived from the multiqc path provided. Used in conjunction with \code{\link[=run_nf_core_rnaseq]{run_nf_core_rnaseq()}}.
}
\examples{
\dontrun{
multiqc_path <- "path_to_a_nf-core_multiqc_report.html"
get_rnaseq_se(multiqc_path)
}
}
\seealso{
\code{\link{run_nf_core_rnaseq}}
}
\author{
Denis O'Meally
}
