% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{make_seurat_objects}
\alias{make_seurat_objects}
\alias{seurat_import_objects}
\title{Make a Seurat object from 10X Cellranger CellPlex data}
\usage{
seurat_import_objects(project_regex)
}
\arguments{
\item{project_regex}{string, optional, regex pattern to match in \code{project}. Use this to select different
cohorts.}
}
\value{
a list of Seurat objects
}
\description{
Loads 10X Cellranger CellPlex h5 data, builds Seurat objects and adds the following metatdata columns:
\code{percent_mt}, \code{percent_ribo}, \code{percent_hb}, \code{percent_platelet}, \code{percent_xist}, \code{chrY_counts}, \code{percent_myh11}.

Data are read in from the "cellranger" directory \verb{/net/nfs-irwrsrchnas01/labs/rrockne/MHO/AML.scRNA.2022/cellranger}
by searching for files named \code{sample_feature_bc_matrix.h5}. This can take some time, so after the 1st run,
the paths are written to \code{data-raw/cellranger_h5_paths.txt} and read from there for subsequent runs.
Similarly, sex chromosome genes are parsed from the GTF and cached in \code{inst/extdata}.

TODO: Include eg for accessing Y genes from package
}
\author{
Denis O'Meally
}
