% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metadata_mmu}
\alias{metadata_mmu}
\title{Minimal metadata for mouse samples}
\source{
\code{\link[=make_metadata_mmu]{make_metadata_mmu()}}
}
\description{
\tabular{ll}{
    \tab Description \cr
   \code{sample} \tab Integrative Genomics Core \code{library_id} with the format \code{COHP_XXXXX}. We use this unique identifier to match sequence reads, QC metrics, raw counts and metadata, for each sample \cr
   \code{fastq_1} \tab Full path of fastq1 \cr
   \code{fastq_2} \tab Full path of fastq2, or blank for single-end reads \cr
   \code{strandedness} \tab Sequencing library protocol (\code{reverse} or \code{unstranded}) \cr
   \code{mouse_id} \tab A unique identifier for each mouse, a 3 or 4 digit number; database held by Kuo lab \cr
   \code{tissue} \tab Tissue type: \code{PBMC} peripheral blood mononuclear cells; \code{BM} bone marrow aspirant; \code{BM_CKIT} ckit+ flow sorted \code{BM} \cr
   \code{weeks} \tab Timepoint in weeks \cr
   \code{timepoint} \tab Timepoint label; formatting is inconsistent across cohorts \cr
   \code{batch} \tab Sequencing groups \cr
   \code{treatment} \tab Experimental treatment group \cr
   \code{genotype} \tab Mouse genotype \cr
   \code{sex} \tab Sex of the mouse \cr
   \code{dob} \tab Date of birth of the mouse \cr
   \code{project} \tab Indicates which samples were processed together; a project may contain multiple \code{batches} \cr
}
}
\details{
The \code{\link[=make_metadata_mmu]{make_metadata_mmu()}} function assembles the metadata for all RNAseq libraries from AML and CML mice.
Minimal metadata fields include sample, fastq_1, fastq_2, strandedness,
mouse_id, tissue, weeks, timepoint, batch, treatment, genotype, sex, dob, project.

Raise an \href{https://github.com/drejom/haemdata/issues}{issue on GitHub}
to report erroneous or missing records.
}
\author{
Denis O'Meally
}
