% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metadata_mmu}
\alias{metadata_mmu}
\title{Minimal metadata for mouse samples}
\source{
\href{https://github.com/drejom/haemdata/blob/main/scripts/import_metadata.R#L24}{\code{update_metadata_mmu()}}
}
\description{
\href{../metadata_mmu_template.xlsx}{Download} a template for new samples\tabular{ll}{
   Column \tab Description \cr
   \code{sample_id} \tab A unique sample identifier with the format \code{PSON_XXXX} that uniquely identifies a tissue sample with respect to the mouse, tissue type and sample date. Use this number to match samples across assays (eg mRNA & miRNA.) \cr
   \code{library_id} \tab Integrative Genomics Core \emph{CLN} with the format \code{COHP_XXXXX}. This identifier is used to match sequence reads, QC metrics, raw counts and metadata, for each sequencing library. \code{library_id} is not unique as some libraries are split across multiple sequencing runs. \cr
   \code{fastq_1} \tab Full path of fastq1, unique in the table \cr
   \code{fastq_2} \tab Full path of fastq2, unique in the table, or blank for single-end reads \cr
   \code{hdf5} \tab Full path of h5 file, Cell Rangers's HDF5 Feature-Barcode Matrix Format \cr
   \code{strandedness} \tab Sequencing library protocol (\code{reverse}, \code{forward} or \code{unstranded}) \cr
   \code{assay} \tab one of \code{mRNA}, \code{miRNA}, \code{scRNA} \cr
   \code{mouse_id} \tab A unique 3 or 4 digit identifier for each mouse; database held by \href{mailto:YKuo@coh.org?subject=Question\%20about\%20AML\%20mice\%20from\%20PSON}{Kuo lab} for AML mice and \href{mailto:YKuo@coh.org?subject=Question\%20about\%20AML\%20mice\%20from\%20PSON}{Zhang lab} for CML. \cr
   \code{tissue} \tab Tissue type: \code{PBMC} peripheral blood mononuclear cells; \code{BM} bone marrow aspirant; \code{BM_CKIT} ckit+ flow sorted \code{BM}; \code{PBMC_CKIT} ckit+ flow sorted \code{PBMC} \cr
   \code{timepoint} \tab Timepoint label; formatting is inconsistent across cohorts \cr
   \code{treatment} \tab Experimental treatment group \cr
   \code{genotype} \tab Mouse genotype \cr
   \code{sex} \tab Sex, \code{M} or \code{F} \cr
   \code{dob} \tab Date of birth of the mouse, YYYY-MM-DD \cr
   \code{project} \tab Cohorts of mice collected and processed together; a project may contain multiple \code{batches} \cr
   \code{batch} \tab Groups of samples sequenced together \cr
   \code{dod} \tab Date of death of the mouse, YYYY-MM-DD. \code{NA} if the mouse survived the experiment. \cr
   \code{sample_date} \tab Date the sample was collected, YYYY-MM-DD \cr
   \code{percent_ckit} \tab Percentage of c-KIT+ cells, measured by flow cytometry (CD117) \cr
   \code{sample_weeks} \tab Timepoint in weeks (\code{sample_date - min(sample_date)}). Post treatment chemo samples begin at week 0, pretreatment samples < 0. \cr
   \code{age_at_start} \tab Age at start of the experiment, in weeks (\code{min(sample_date) - dob}) \cr
   \code{age_at_sample} \tab Age at sample collection, in weeks (\code{sample_date - dob}) \cr
   \code{age_at_end} \tab Age at end of of the experiment, in weeks (\code{max(sample_date) - dob}) \cr
   \code{qc_pass_mapping} \tab \code{mRNA} samples only. \code{TRUE} if STAR uniquely mapped reads >= \code{mapping_threshold} (5\% by default), or \code{FALSE} if not. \cr
   \code{ref_dim1} \tab reference dimension 1 - typically PC1; can be UMAP or any other dimension reduction coordinate. \cr
   \code{ref_dim1} \tab reference dimension 2 - typically PC2 \cr
}
}
\details{
The \href{https://github.com/drejom/haemdata/blob/main/scripts/import_metadata.R#L24}{\code{update_metadata_mmu()}}
function assembles the metadata for all RNAseq libraries from AML and CML mice, by consolidating
data scraped from multiple sequencing run sheets, directly from sequencing folders, emails, and so forth.
The code is complex and ugly and undoubtedly some errors will have made it through.

Raise an \href{https://github.com/drejom/haemdata/issues}{issue on GitHub}
to report erroneous or missing records.
}
\author{
Denis O'Meally
}
