% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publish_data.R
\name{publish_seurat}
\alias{publish_seurat}
\title{Publish a Seurat object}
\usage{
publish_seurat(seurat_object)
}
\arguments{
\item{seurat_object}{a Seurat object ready to publish}
}
\value{
the path to the published object
}
\description{
Publishes a Seurat object to the Haemdata Teams channel or to the \code{MHO/haemdata} folder
on the COH Isilon storage (in \code{.rda} format) along with a ScanPy object (in \code{.h5ad} format).
Files are named according to the Seurat object's \code{name} slot.
}
\author{
Denis O'Meally
}
