% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{published_pins}
\alias{published_pins}
\title{Datasets published to the Haemdata pin board}
\format{
A \code{data.frame} with the following columns:
\describe{
\item{pin_name}{Name of the dataset on the pin board, including the file extension}
\item{version}{Version of the pin to which the dataset belongs}
}
}
\usage{
data(published_pins)
}
\description{
Lists the name and version of each dataset published to the Haemdata pinboard in the current release.
Pins can be retrieved with the \code{\link[=get_pin]{get_pin()}} function.
}
\examples{
head(published_pins, n = Inf)
}
