% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq.R, R/smrnaseq.R
\name{run_nf_core_rnaseq}
\alias{run_nf_core_rnaseq}
\alias{run_nfcore_smrnaseq}
\alias{run_cutadapt_smrna}
\title{Run nf-core/rnaseq pipeline on Apollo}
\usage{
run_nf_core_rnaseq(run_folder, sample_sheet, ref_genome, qc = TRUE)

run_nfcore_smrnaseq(
  run_folder,
  sample_sheet,
  clip_r1_3p = NULL,
  species = "mmu"
)

run_cutadapt_smrna(sample_sheet)
}
\arguments{
\item{run_folder}{subfolder in which to run nextflow, keeps the pipeline
from overwriting previous runs and enables the \code{-resume} feature of Nextflow.}

\item{sample_sheet}{nf-core sample sheet with the columns \verb{sample_id, library_id, fastq_1}}

\item{ref_genome}{nf-core reference genome, see \href{../articles/genomes.html}{Reference Genomes} for supported references}

\item{qc}{run the full QC pipeline with STAR_Salmon, or Salmon pseudoalignment only, skipping all QC steps (default: TRUE)}

\item{clip_r1_3p}{number of bases to trim from reads.}

\item{species}{mmu or hsa. If \emph{mmu}, additionally sets the 3' adapter to TCTGGAATTCTCGGGTGCCAAGGAACTCC. For \emph{hsa}, the 3' adapter is auto-discovered.}
}
\value{
a path to the run script, or the multiqc report if it exists.

a path to the run script, or the multiqc report if it exists.

a data.frame with the columns \verb{library_id, fastq_1}, ready to be used as input to the nf-core pipeline
}
\description{
Makes an \code{sbatch} job script for the nf-core/rnaseq pipeline and submits it to the Apollo cluster.
The pipeline version is specified by the \code{rnaseq_release} parameter in the \verb{_targets.R} file. The \href{https://nf-co.re/rnaseq/3.7/usage}{default parameters} are used for the pipeline, in addition to \verb{--pseudo_aligner salmon} and \code{--deseq2_vst}.

Makes an \code{sbatch} job script for the nf-core/smrnaseq pipeline and submits it to the Apollo cluster.
The pipeline version is specified by the \code{smrnaseq_release} parameter in the \verb{_targets.R} file. The \href{https://nf-co.re/smrnaseq/2.1.0/usage}{default parameters} are used for the pipeline.

Makes an \code{sbatch} job script for cutadapt and submits it to the Apollo cluster, one job per library.
}
\details{
Execution of the pipeline is independent of the package build script, allowing asynchronous development.
If no \code{run_folder} exists, it will be created along with a \code{sample_sheet.csv} and \verb{run_\{ref_genome\}.sh},
which is submitted via \code{sbatch}. A successful run is detected by the presence of a a multiqc report.
An existing run script wont be overwritten by this function so it must be deleted manually.

Execution of the pipeline is independent of the package build script, allowing asynchronous development.
If no \code{run_folder} exists, it will be created along with a \code{sample_sheet.csv} and \verb{run_\{ref_genome\}.sh},
which is submitted via \code{sbatch}. A successful run is detected by the presence of a a multiqc report.
An existing run script wont be overwritten by this function so it must be deleted manually.

Execution of the pipeline is independent of the package build script, allowing asynchronous development.
If no \code{run_folder} exists, it will be created along with a \code{sample_sheet.csv} and \verb{run_\{ref_genome\}.sh},
which is submitted via \code{sbatch}. A successful run is detected by the presence of a a multiqc report.
An existing run script wont be overwritten by this function so it must be deleted manually.
}
\author{
Denis O'Meally
}
