% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{seurat_annotate_cell_cycle}
\alias{seurat_annotate_cell_cycle}
\title{Annotate mouse cell cycle}
\source{
https://www.science.org/doi/abs/10.1126/science.aad0501
}
\usage{
seurat_annotate_cell_cycle(seurat_object)
}
\arguments{
\item{seurat_object}{A Seurat object to annotate}
}
\value{
A Surat object with cell cycle annotations added to cell \code{meta.data} (\code{S.Score}, \code{G2M.Score} & \code{Phase})
}
\description{
Uses the function \code{\link[Seurat:CellCycleScoring]{Seurat::CellCycleScoring()}} to annotate the cells with their cell cycle \code{Phase},
, \code{S.score} and \code{G2M.score}. The Seurat function uses an internal list of human gene symbols from
Tirosh et al 2016.
#'
Here we substitute equivalent mouse symbols from \code{biomaRt},
but the service was down at time of writing. In a quick hack, we simply
convert gene symbols by making them lowercase, with the 1st letter in caps.
}
\details{
https://rdrr.io/cran/Seurat/man/CellCycleScoring.html
https://github.com/satijalab/seurat/issues/2493
}
\author{
Denis O'Meally
}
