% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_annotate_clusters_and_umap.R
\name{seurat_annotate_umap_and_clusters}
\alias{seurat_annotate_umap_and_clusters}
\alias{seurat_annotate_clusters_and_umap}
\title{Annotate UMAP and knn clusters}
\usage{
seurat_annotate_clusters_and_umap(seurat_object)
}
\arguments{
\item{seurat_object}{a Seurat object}
}
\value{
a Seurat object with UMAP and knn clusters
}
\description{
Runs the 'standard workflow' using 2000 features, but differs in that the number of PCs used for clustering is
is first determined using using the \code{\link[=find_elbow]{find_elbow()}} function. The number of clusters is
tuned using the \code{cluster_res} parameter within the \code{\link[Seurat:FindClusters]{Seurat::FindClusters()}} function.
}
\details{
https://satijalab.org/seurat/articles/integration_introduction.html#perform-an-integrated-analysis-1
}
\author{
Denis O'Meally
}
