% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{seurat_perform_cell_qc}
\alias{seurat_perform_cell_qc}
\title{Basic single cell QC}
\usage{
seurat_perform_cell_qc(raw_seurat_objects)
}
\arguments{
\item{raw_seurat_objects}{a list of Seurat objects, fresh in from Cellranger}
}
\value{
a list of Seurat objects
}
\description{
Applies the following filters to the data:
\enumerate{
\item Removes genes expressed in fewer than 3 cells
\item Removes cells with fewer than 200 genes expressed
\item Removes cells where mitochondrial genes account for more than 20\% of expressed transcripts
\item Removes cells where ribosomal genes account for fewer than 5\% of expressed transcripts
}
}
\author{
Denis O'Meally
}
