% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_data.R
\name{use_pinboard}
\alias{use_pinboard}
\title{Setup the pinboard for use}
\usage{
use_pinboard(
  pin_board = NULL,
  haemdata_folder = "/net/isi-dcnl/ifs/user_data/rrockne/MHO/haemdata"
)
}
\arguments{
\item{pin_board}{Indicate which pinboard to use. One of c('onedrive,
'isilon'), Default: NULL}

\item{haemdata_folder}{The full path of the Isilon folder holding Haemdata
pins, Default: '/net/isi-dcnl/ifs/user_data/rrockne/MHO/haemdata'}
}
\value{
Nothing. Assigns the \code{pin_board} global variable accordingly
}
\description{
Sets up the pinboard to use for the current R session.
}
\details{
Package releases always use the OneDrive pinboard with versioned pins.
Pin versions can be retrieved from the \code{published_pins} variable.
Development versions of the package use the Isilon pinboard with unversioned pins.
Only the most recent pin is kept.
}
\examples{
\dontrun{
if (interactive()) {
    # Use the OneDrive pinboard (default)
    use_pinboard("onedrive")
    # Use the Isilon pinboard
    use_pinboard("isilon")
}
}
}
\seealso{
\code{\link[pins]{board_ms365}}, \code{\link[pins]{board_folder}}
\code{\link[Microsoft365R]{get_personal_onedrive}}
}
\author{
Denis O'Meally
}
