% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_data.R
\name{use_pinboard}
\alias{use_pinboard}
\title{Setup the pinboard for use}
\usage{
use_pinboard(
  pin_board = NULL,
  haemdata_folder = "/net/nfs-irwrsrchnas01/labs/rrockne/MHO/haemdata"
)
}
\arguments{
\item{pin_board}{Indicate which pinboard to use. One of c('onedrive,
'devel'), Default: NULL}

\item{haemdata_folder}{The full path of the folder for devel
pins, Default: '/net/nfs-irwrsrchnas01/labs/rrockne/MHO/haemdata'}
}
\value{
Assigns the \code{pin_board} variable in the haemdata_env environment
}
\description{
Sets up the pinboard to use for the current R session. Rather
than issuing this command each session, you can add the line
\code{if (require(haemdata)) haemdata::use_pinboard("onedrive")} to your project folder
\href{https://support.rstudio.com/hc/en-us/articles/360047157094-Managing-R-with-Rprofile-Renviron-Rprofile-site-Renviron-site-rsession-conf-and-repos-conf}{\code{.Rprofile}} file.
}
\details{
Package releases always use the OneDrive pinboard with versioned pins.
Pin versions associated with each release can be retrieved with \code{data(published_pins)}.
Development pins are published to the \code{haemdata_folder}, but are not versioned.
}
\examples{
\dontrun{
if (interactive()) {
    # Use the OneDrive pinboard
    use_pinboard("onedrive")
    # Use the Development pinboard
    use_pinboard("devel")
}
}
}
\seealso{
\code{\link[pins]{board_ms365}}, \code{\link[pins]{board_folder}}
\code{\link[Microsoft365R]{get_personal_onedrive}}
}
\author{
Denis O'Meally
}
